remove intellij-idea.topic and replace references with new topic
