--- codeSnippets/snippets/server-get-started/src/main/kotlin/com/example/Application.kt
@@ -1,15 +1,0 @@
-package com.example
-
-import com.example.plugins.*
-import io.ktor.server.application.*
-import io.ktor.server.engine.*
-import io.ktor.server.netty.*
-
-fun main() {
-    embeddedServer(Netty, port = 9292, host = "0.0.0.0", module = Application::module)
-        .start(wait = true)
-}
-
-fun Application.module() {
-    configureRouting()
-}
