--- codeSnippets/snippets/tutorial-server-get-started/README.md
@@ -1,13 +1,0 @@
-# Getting started with a Ktor Server
-
-A sample project created in the [Creating a new Ktor project](https://ktor.io/docs/intellij-idea.html) tutorial.
-> This sample is a part of the [codeSnippets](../../README.md) Gradle project.
-
-## Running
-
-To run a sample, execute the following command in a repository's root directory:
-
-```bash
-./gradlew :tutorial-server-get-started:run
-```
-
