--- codeSnippets/snippets/tutorial-server-get-started/build.gradle.kts
@@ -1,31 +1,0 @@
-val ktor_version: String by project
-val kotlin_version: String by project
-val logback_version: String by project
-
-plugins {
-    application
-    kotlin("jvm")
-    id("io.ktor.plugin") version "2.3.7"
-}
-
-group = "com.example"
-version = "0.0.1"
-application {
-    mainClass.set("com.example.ApplicationKt")
-
-    val isDevelopment: Boolean = project.ext.has("development")
-    applicationDefaultJvmArgs = listOf("-Dio.ktor.development=$isDevelopment")
-}
-
-repositories {
-    mavenCentral()
-    maven { url = uri("https://maven.pkg.jetbrains.space/public/p/ktor/eap") }
-}
-
-dependencies {
-    implementation("io.ktor:ktor-server-core-jvm:$ktor_version")
-    implementation("io.ktor:ktor-server-netty-jvm:$ktor_version")
-    implementation("ch.qos.logback:logback-classic:$logback_version")
-    testImplementation("io.ktor:ktor-server-tests-jvm:$ktor_version")
-    testImplementation("org.jetbrains.kotlin:kotlin-test-junit:$kotlin_version")
-}
