+++ codeSnippets/snippets/tutorial-server-get-started/src/main/kotlin/com/example/Application.kt
@@ -1,12 +1,12 @@
 package com.example
 
+import com.example.plugins.*
 import io.ktor.server.application.*
 import io.ktor.server.engine.*
 import io.ktor.server.netty.*
-import com.example.plugins.*
 
 fun main() {
-    embeddedServer(Netty, port = 8080, host = "0.0.0.0", module = Application::module)
+    embeddedServer(Netty, port = 9292, host = "0.0.0.0", module = Application::module)
         .start(wait = true)
 }
