+++ codeSnippets/snippets/tutorial-server-get-started/src/main/resources/mycontent/sample.html
