+++ codeSnippets/snippets/tutorial-server-get-started/src/test/kotlin/com/example/ApplicationTest.kt
@@ -1,25 +1,32 @@
 package com.example
 
-import io.ktor.server.routing.*
-import io.ktor.http.*
-import io.ktor.server.application.*
-import io.ktor.server.response.*
-import io.ktor.server.request.*
 import io.ktor.client.request.*
 import io.ktor.client.statement.*
-import kotlin.test.*
+import io.ktor.http.*
 import io.ktor.server.testing.*
-import com.example.plugins.*
+import org.junit.Assert.assertEquals
+import org.junit.Test
+import kotlin.test.assertContains
 
 class ApplicationTest {
     @Test
     fun testRoot() = testApplication {
         application {
-            configureRouting()
+            module()
         }
-        client.get("/").apply {
-            assertEquals(HttpStatusCode.OK, status)
-            assertEquals("Hello World!", bodyAsText())
+        val response = client.get("/")
+        assertEquals(HttpStatusCode.OK, response.status)
+        assertEquals("Hello World!", response.bodyAsText())
+    }
+    @Test
+    fun testNewEndpoint() = testApplication {
+        application {
+            module()
         }
+
+        val response = client.get("/test1")
+        assertEquals(HttpStatusCode.OK, response.status)
+        assertEquals("html", response.contentType()?.contentSubtype)
+        assertContains(response.bodyAsText(), "Hello From Ktor")
     }
 }
