+++ topics/FAQ.topic
@@ -85,8 +85,12 @@
     <chapter title="How to resolve 'No configuration setting found for key ktor'?" id="cannot-find-application-conf" level="3">
         <p>
             This means that Ktor was not able to find a <a href="Configuration-file.topic">configuration file</a>.
-            Re-check that it is in the <code>resources</code> folder, and that the <code>resources</code> folder is marked as such.
-            You can consider to set up a project using the <a href="https://start.ktor.io/">project generator</a> or the <a href="intellij-idea.topic">IntelliJ plugin</a> to have a working project as the base.
+            Ensure that there is a configuration file in the <code>resources</code> folder, and that the
+            <code>resources</code> folder is marked as such.
+            Consider setting up a project using the <a href="https://start.ktor.io/">Ktor project generator</a>
+            or the
+            <a href="https://plugins.jetbrains.com/plugin/16008-ktor">Ktor plugin for intelliJ IDEA Ultimate</a> to have
+            a working project as the base. For more information, see <a href="server_create_a_new_project.topic"/>.
         </p>
     </chapter>
