+++ topics/welcome.topic
@@ -12,7 +12,7 @@
         </description>
 
         <spotlight>
-            <a href="intellij-idea.topic" type="server"
+            <a href="server_create_a_new_project.topic" type="server"
                summary="Learn how to create, run and test a server application with Ktor.">
                 Get started with Ktor Server
             </a>
@@ -42,7 +42,7 @@
             <links narrow="true">
                 <group>
                     <title>Server configuration</title>
-                    <a href="intellij-idea.topic"/>
+                    <a href="server_create_a_new_project.topic"/>
                     <a href="server-dependencies.topic"/>
                     <a href="create_server.topic"/>
                     <a href="Configuration-code.topic"/>
